# knowledge-dao-server/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Knowledge DAO API
  description: |
    本地 RAG 知识库 API，支持：
    - **RAG 核心**：语义检索、知识插入、RAG 问答
    - **管理后台**：知识管理、用户会话、系统监控、配置管理
    - **仪表盘**：驾驶舱统计数据
    - **检索分析**：搜索历史、热词统计

    基础路径：
    - ApiServer（用户 API）：`http://localhost:8080`
    - AdminApiServer（管理 API）：`http://localhost:8081`
  version: "1.0.0"
  contact:
    name: knowledge-dao team
  license:
    name: MIT
    url: https://github.com/chqiuu/knowledge-dao
servers:
  - url: http://localhost:8080
    description: 用户 API（ApiServer）
  - url: http://localhost:8081
    description: 管理后台 API（AdminApiServer）
...
paths:
...
  /api/admin/entries:
    get:
      operationId: admin-entries-list
      summary: 知识条目列表（分页）
      tags: [管理后台-知识管理]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
        - name: tag
          in: query
          schema:
            type: string
        - name: keyword
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 列表成功
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
    post:
      operationId: admin-entries-create
      summary: 创建知识条目
      tags: [管理后台-知识管理]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertRequest"
      responses:
        "201":
          description: 创建成功
        "400":
          description: 参数校验失败
  /api/admin/entries/{id}:
    get:
      operationId: admin-entries-get
      summary: 知识条目详情
      tags: [管理后台-知识管理]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
        "404":
          description: 条目不存在
    put:
      operationId: admin-entries-update
      summary: 更新知识条目
      tags: [管理后台-知识管理]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InsertRequest"
      responses:
        "200":
          description: 更新成功
        "404":
          description: 条目不存在
    delete:
      operationId: admin-entries-delete
      summary: 删除知识条目
      tags: [管理后台-知识管理]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 删除成功
        "404":
          description: 条目不存在
  /api/admin/entries/batch:
    post:
      operationId: admin-entries-batch
      summary: 批量导入知识
      tags: [管理后台-知识管理]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BatchImportRequest"
      responses:
        "200":
          description: 导入完成
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /api/admin/entries/{id}/rebuild:
    post:
      operationId: admin-entries-rebuild
      summary: 重建向量
      description: 重新生成指定知识条目的向量
      tags: [管理后台-知识管理]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 重建成功
        "404":
          description: 条目不存在
  /api/admin/searches:
    get:
      operationId: admin-searches-list
      summary: 搜索历史列表
      tags: [管理后台-搜索分析]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
...
  /api/admin/users:
    get:
      operationId: admin-users-list
      summary: 用户列表
      tags: [管理后台-用户会话]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: 成功
  /api/admin/sessions:
    get:
      operationId: admin-sessions-list
      summary: 会话列表
      tags: [管理后台-用户会话]
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: userId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: 成功
...
components:
  schemas:
    ApiResponse:
      type: object
      description: 统一响应包装（对齐 kin-tree-cloud Result<T>）
      properties:
        code:
          type: integer
          description: "状态码：1=成功，其他=失败"
          example: 1
        message:
          type: string
          description: 状态描述
          example: "success"
        data:
          type: object
          nullable: true
          description: 响应数据
        time:
          type: integer
          format: int64
          description: 服务器时间戳（毫秒）
          example: 1742544000000
    InsertRequest:
      type: object
      required:
        - title
        - content
        - userId
      properties:
        title:
          type: string
          maxLength: 500
          description: 标题
          example: "Java 并发编程"
        content:
          type: string
          description: 正文内容
        contentType:
          type: string
          description: 内容类型
          default: "article"
          enum: [article, tutorial, book, note, doc, qa]
        tags:
          type: array
          items:
            type: string
          maxItems: 20
          description: 标签列表
          example: ["java", "concurrency"]
        source:
          type: string
          description: 来源标识
          default: "api"
        userId:
          type: integer
          format: int64
          minimum: 1
          description: 所属用户 ID
...
    BatchImportRequest:
      type: object
      required:
        - entries
      properties:
        entries:
          type: array
          maxItems: 100
          items:
            type: object
            required:
              - title
              - content
              - userId
            properties:
              title:
                type: string
                maxLength: 500
              content:
                type: string
              contentType:
                type: string
                default: "article"
              tags:
                type: array
                items:
                  type: string
                maxItems: 20
              source:
                type: string
                default: "import"
              userId:
                type: integer
                format: int64
                minimum: 1
...